pyproject.toml
[project]
name = "thalassa"
version = "0.1.0"
description = "Local Bruin, DuckDB, and Streamlit project for Greek sailing traffic analytics."
requires-python = ">=3.11"
dependencies = [
    "db-dtypes>=1.4.3",
    "duckdb>=1.2",
    "google-cloud-bigquery>=3.38.0",
    "jupyter>=1.1.1",
    "pandas>=2.2",
    "pyarrow>=20.0.0",
    "python-dotenv>=1.1.1",
    "pydantic>=2.11.7",
    "requests>=2.32",
    "streamlit>=1.44",
    "streamlit-option-menu>=0.4.0",
]
